+++ apps/neon-cinematic/src/App.tsx
@@ -1,0 +1,7 @@
+import { NeonCinematicDemo } from "./components/NeonCinematicDemo";
+
+function App() {
+  return <NeonCinematicDemo />;
+}
+
+export default App;